Rename Redux based API functions
