+++ src/api/reports/accountSummaryReport.test.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { runReport } from './accountSummaryReports';
+import { runReport } from './accountSummaryReport';
 import { ReportType } from './report';
 
 test('api run reports calls axios get', () => {
