+++ src/api/reports/detailsReport.ts
