+++ src/routes/overview/components/committed-spend/CommittedSpend.tsx
@@ -1,5 +1,5 @@
 import { getQuery } from 'api/queries';
-import type { AccountSummaryReport } from 'api/reports/accountSummaryReports';
+import type { AccountSummaryReport } from 'api/reports/accountSummaryReport';
 import type { AxiosError } from 'axios';
 import messages from 'locales/messages';
 import React, { useEffect } from 'react';
